logic/Detect/AIScanBIAS/requirements.txt
fastapi
uvicorn[standard]
apscheduler
openai
python-dotenv
pytz ; python_version < "3.9"
